# Add support and switch between of rsa2048, rsa4096 and ed25519 keys