--- a/src/crypto.h
+++ b/src/crypto.h
@@ -89,12 +89,17 @@ class Crypto {
  public:
   static std::string sha256digest(const std::string &text);
   static std::string sha512digest(const std::string &text);
-  static std::string RSAPSSSign(ENGINE *engine, const std::string &private_key, const std::string &message);
+  static std::string RSAPSSSign(ENGINE *engine, const std::string &private_key, const std::string &digest);
+  static std::string Sign(KeyType key_type, ENGINE *engine, const std::string &private_key, const std::string &digest);
+  static std::string ED25519Sign(const std::string &private_key, const std::string &digest);
+
   static bool VerifySignature(const PublicKey &public_key, const std::string &signature, const std::string &message);
   static bool parseP12(FILE *p12_fp, const std::string &p12_password, std::string *out_pkey, std::string *out_cert,
                        std::string *out_ca);
   static bool extractSubjectCN(const std::string &cert, std::string *cn);
-  static bool generateRSAKeyPair(std::string *public_key, std::string *private_key);
+  static bool generateRSAKeyPair(KeyType key_type, std::string *public_key, std::string *private_key);
+  static bool generateEDKeyPair(std::string *public_key, std::string *private_key);
+  static bool generateKeyPair(KeyType key_type, std::string *public_key, std::string *private_key);
 
   static bool RSAPSSVerify(const std::string &public_key, const std::string &signature, const std::string &message);
   static bool ED25519Verify(const std::string &public_key, const std::string &signature, const std::string &message);