--- a/src/uptane/managedsecondary.cc
+++ b/src/uptane/managedsecondary.cc
@@ -14,7 +14,8 @@ ManagedSecondary::ManagedSecondary(const SecondaryConfig &sconfig_in) : Secondar
   // loadMetadata(meta_pack);
   if (!loadKeys(&public_key, &private_key)) {
     if (!Crypto::generateKeyPair(sconfig.key_type, &public_key, &private_key)) {
-      LOG_ERROR << "Could not generate rsa keys for secondary " << getSerial() << "@" << sconfig.ecu_hardware_id;
+      LOG_ERROR << "Could not generate rsa keys for secondary " << ManagedSecondary::getSerial() << "@"
+                << sconfig.ecu_hardware_id;
       throw std::runtime_error("Unable to generate secondary rsa keys");
     }
     storeKeys(public_key, private_key);