--- a/tests/config_tests_device_id.toml
+++ b/tests/config_tests_device_id.toml
@@ -11,6 +11,7 @@ websocket_server = "127.0.0.1:3012"
 
 [uptane]
 device_id = "test-name-123"
+key_type = "ED25519"
 
 [provision]
 provision_path = "tests/test_data/cred.zip"