--- a/tests/config_tests_prov.toml
+++ b/tests/config_tests_prov.toml
@@ -19,6 +19,7 @@ server = "https://7d0a4914-c392-4ccd-a8f9-3d4ed969da07.tcpgw.prod01.advancedtele
 [uptane]
 polling = false
 polling_sec = 91
+key_type = "ED25519"
 
 [pacman]
 type = "none"