Running all uptane related tests with command `ctest -R test_uptane` gives me the following results:
```
RSA2048 : 120 secs
ED255519: 60 secs
```
So looks like ED25519 is much faster then RSA2048. 
But server doesnt support it, so we can not use it by default yet.