@cajun-rat 2018-02-16
Let's hold off merging this until the Debian stuff is done and we have a working release of that.  I don't want to merge anything risky until we have this out the door.