@patriotyk 2018-02-21
Looks like there is no any benefits in speedin CI. So I will update PR with default ED keys for all tests to check if something change