@patriotyk 2018-02-21
I think it is ok now. However there is no any speedup profit comparing to RSA2048 even when run all tests with ED25519. But on my local machine ED25519 performs much faster