@cajun-rat 2018-02-22
Benchmark: 

## With these changes (3 runs)

~~~
ninja clean && time ninja qa

Total Test time (real) = 189.70 sec
real	4m26,985s
user	15m30,724s
sys	0m33,272s

Total Test time (real) = 170.85 sec
real	4m8,790s
user	15m10,832s
sys	0m34,140s

Total Test time (real) = 185.23 sec
real	4m23,467s
user	15m23,628s
sys	0m34,220s

~~~

## without (master): 

~~~
Total Test time (real) = 191.73 sec
real	4m29,087s
user	15m26,884s
sys	0m33,652s

Total Test time (real) = 198.52 sec
real	4m36,718s
user	15m38,312s
sys	0m34,048s

Total Test time (real) = 207.30 sec
real	4m45,438s
user	15m34,548s
sys	0m34,204s
~~~